requirements.txt
amqp==5.1.1
asgiref==3.2.10
async-timeout==4.0.2
beautifulsoup4==4.9.1
billiard==3.6.4.0
bs4==0.0.1
celery==5.2.7
certifi==2020.6.20
cffi==1.15.0
chardet==3.0.4
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cryptography==37.0.4
defusedxml==0.6.0
dj-database-url==0.5.0
Django==3.1.11
django-allauth==0.42.0
django-celery-beat==2.4.0
django-celery-results==2.4.0
django-cors-headers==3.5.0
django-countries==6.1.3
django-filter==2.2.0
django-heroku==0.3.1
django-rest-auth==0.9.5
django-rest-knox==4.1.0
django-rest-passwordreset==1.1.0
django-social-share==2.0.0
django-timezone-field==5.0
djangorestframework==3.11.0
djangorestframework-simplejwt==5.2.2
emoji==2.0.0
gunicorn==20.0.4
html5lib==1.1
idna==2.10
importlib-metadata==1.6.0
kombu==5.2.4
lxml==4.9.1
Markdown==3.2.2
oauthlib==3.1.0
packaging==21.3
Pillow==9.2.0
prompt-toolkit==3.0.33
psycopg2==2.9.3
psycopg2-binary==2.9.1
pycparser==2.20
PyJWT==1.7.1
pyngrok==5.1.0
pyparsing==3.0.9
python-crontab==2.6.0
python-dateutil==2.8.2
python-http-client==3.3.1
python3-openid==3.2.0
pytz==2022.6
PyYAML==6.0
redis==4.3.5
requests==2.24.0
requests-oauthlib==1.3.0
sendgrid==6.9.7
six==1.14.0
soupsieve==2.0.1
sqlparse==0.3.1
starkbank-ecdsa==2.2.0
stripe==2.50.0
twilio==6.51.0
tzdata==2022.6
urllib3==1.25.10
vine==5.0.0
wcwidth==0.2.5
webencodings==0.5.1
whitenoise==5.0.1
zipp==3.1.0
